Add justificationa and fix truncation bugs
This renames left_trim_pad/2 to fit_to_width/3 to better describe what
it does now that it handles right and center justification. This is
backwards incompatible.

The previous hack of adding `\b` characters to trim strings was removed.
In addition to being ugly, it did the wrong thing on some inputs. This
also adds unit tests to catch these cases in the future.