Pull padding and align internal and support multi-line cells
Multi-line cells introduced a challenge in that cells need to be sliced.
This means that Tablet needs to style each slice of a line and the step
to pad/align needs to happen before the styling callbacks. This ends up
simplifying the styling function more which was nice. None of the
options do anything interesting with text alignment, but if this is
needed, some way to specify it will be needed.